# docs/swagger.yaml
basePath: /api
definitions:
  damagerequest.DamageRequest:
    properties:
      ap:
        description: Armor penetration
        type: integer
      attacks_string:
        description: e.g., "D6+2"
        type: string
      bs:
        description: Ballistic Skill
        type: integer
      d:
        description: Damage, e.g., "D3", "2"
        type: string
      devastating_wounds:
        type: boolean
      feel_no_pain:
        description: Feel No Pain, optional
        type: integer
      hit_modifier:
        type: integer
      hit_reroll:
        $ref: '#/definitions/damagerequest.RerollType'
      invulnerable:
        description: Pointers (*int) are used for optional fields. If the field is
          omitted in JSON, the pointer will be nil.
        type: integer
      lethal_hits:
        type: boolean
      num_models:
        type: integer
      s:
        description: Strength
        type: integer
      save:
        description: Target Save
        type: integer
      save_modifier:
        type: integer
      t:
        description: Target Toughness
        type: integer
      torrent:
        type: boolean
      wound_modifier:
        type: integer
      wound_reroll:
        $ref: '#/definitions/damagerequest.RerollType'
    type: object
  damagerequest.DamageResponse:
    properties:
      average_destroyed:
        type: number
      average_hits:
        type: number
      destroyed_distribution:
        additionalProperties:
          format: float64
          type: number
        type: object
      hits_distribution:
        additionalProperties:
          format: float64
          type: number
        type: object
      message:
        type: string
      pens_distribution:
        additionalProperties:
          format: float64
          type: number
        type: object
      wounds_distribution:
        additionalProperties:
          format: float64
          type: number
        type: object
    type: object
  damagerequest.RerollType:
    enum:
    - 0
    - 1
    - 2
    type: integer
    x-enum-varnames:
    - RerollNone
    - RerollOnes
    - RerollFail
host: localhost:8080
info:
  contact: {}
  description: API for calculating damage statistics based on 10th Edition rules.
  title: Warhammer 40k 10th Calc API
  version: "1.0"
paths:
  /damage/calculate:
    post:
      consumes:
      - application/json
      description: Calculates statistical damage based on input parameters
      parameters:
      - description: Calculation Parameters
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/damagerequest.DamageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/damagerequest.DamageResponse'
      summary: Calculate Damage
      tags:
      - damage
swagger: "2.0"