# pkg/handler/damage_handler.go
package handler

import (
	"encoding/json"
	"io"
	"log"
	"net/http"

	damagerequest "github.com/AnNoName1/warhammer40k10thCalc/pkg/models"

	calculator "github.com/AnNoName1/warhammer40k10thCalc/internal/calculator"
)

// CalculateDamageHandler calculates the expected damage.
//
//	@Summary		Calculate Damage
//	@Description	Calculates statistical damage based on input parameters
//	@Tags			damage
//	@Accept			json
//	@Produce		json
//	@Param			request	body		damagerequest.DamageRequest	true	"Calculation Parameters"
//	@Success		200		{object}	damagerequest.DamageResponse
//	@Router			/damage/calculate [post]
func CalculateDamageHandler(w http.ResponseWriter, r *http.Request) {
	if r.Method != http.MethodPost {
		http.Error(w, "Method Not Allowed", http.StatusMethodNotAllowed)
		return
	}

	// 1. Decode the Request Body
	body, err := io.ReadAll(r.Body)
	if err != nil {
		http.Error(w, "Failed to read request body", http.StatusInternalServerError)
		return
	}

	var req damagerequest.DamageRequest
	if err := json.Unmarshal(body, &req); err != nil {
		log.Printf("Error decoding JSON: %v", err)
		http.Error(w, "Invalid JSON format: "+err.Error(), http.StatusBadRequest)
		return
	}

	// 2. Call Core Business Logic
	resp, err := calculator.CalculateDamageCore(req)
	if err != nil {
		log.Printf("Calculation error: %v", err)
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	// 3. Encode the Response Body
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusOK)

	if err := json.NewEncoder(w).Encode(resp); err != nil {
		log.Printf("Error encoding JSON response: %v", err)
	}
}